BeeDetectionModel/bee-detection-runpod-final/requirements.txt
python-dotenv==0.19.2
ultralytics==8.3.159
paramiko==2.10.1
python-dotenv==0.19.2
opencv-python==4.11.0.86
numpy==1.26.4